+++ Editor/APIInternal/ComponentInfos.VRCSDK.cs
@@ -3,7 +3,6 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using Anatawa12.AvatarOptimizer.API;
-using Anatawa12.AvatarOptimizer.ErrorReporting;
 using JetBrains.Annotations;
 using UnityEngine;
 using VRC.SDK3;
@@ -62,8 +61,7 @@ namespace Anatawa12.AvatarOptimizer.APIInternal.VRCSDK
                         collector.ModifyProperties(component.VisemeSkinnedMesh,
                             $"blendShape.{component.MouthOpenBlendShapeName}");
                     } else {
-                        BuildReport.LogWarning("ComponentInfos:VRCAvatarDescriptor:warning:NoVisemeSkinnedMesh")
-                            ?.WithContext(component);
+                        BuildLog.LogWarning("ComponentInfos:VRCAvatarDescriptor:warning:NoVisemeSkinnedMesh", component);
                     }
                     break;
                 }
@@ -74,8 +72,7 @@ namespace Anatawa12.AvatarOptimizer.APIInternal.VRCSDK
                         collector.ModifyProperties(component.VisemeSkinnedMesh,
                             component.VisemeBlendShapes.Select(blendShape => $"blendShape.{blendShape}"));
                     } else {
-                        BuildReport.LogWarning("ComponentInfos:VRCAvatarDescriptor:warning:NoVisemeSkinnedMesh")
-                            ?.WithContext(component);
+                        BuildLog.LogWarning("ComponentInfos:VRCAvatarDescriptor:warning:NoVisemeSkinnedMesh", component);
                     }
                     break;
                 }
@@ -83,9 +80,9 @@ namespace Anatawa12.AvatarOptimizer.APIInternal.VRCSDK
                     // NOP
                     break;
                 default:
-                    BuildReport.LogWarning("ComponentInfos:VRCAvatarDescriptor:warning:UnknownLipSyncStyle", 
-                            component.lipSync.ToString())
-                        ?.WithContext(component);
+                    BuildLog.LogWarning("ComponentInfos:VRCAvatarDescriptor:warning:UnknownLipSyncStyle", 
+                            component.lipSync.ToString(),
+                            component);
                     break;
             }
         }
@@ -130,7 +127,7 @@ namespace Anatawa12.AvatarOptimizer.APIInternal.VRCSDK
                             removed = true;
                     }
                     if (removed)
-                        BuildReport.LogFatal("ApplyObjectMapping:VRCAvatarDescriptor:viseme BlendShape Removed");
+                        BuildLog.LogError("ApplyObjectMapping:VRCAvatarDescriptor:viseme BlendShape Removed");
                     break;
                 }
                 case VRC_AvatarDescriptor.LipSyncStyle.VisemeParameterOnly:
@@ -182,9 +179,9 @@ namespace Anatawa12.AvatarOptimizer.APIInternal.VRCSDK
                     case VRCAvatarDescriptor.ColliderConfig.State.Disabled:
                         break;
                     default:
-                        BuildReport.LogWarning("ComponentInfos:VRCAvatarDescriptor:warning:UnknownColliderState",
-                                collider.ToString(), where)
-                            ?.WithContext(component);
+                        BuildLog.LogWarning("ComponentInfos:VRCAvatarDescriptor:warning:UnknownColliderState",
+                                collider.ToString(), where,
+                                component);
                         break;
                 }
             }
@@ -234,21 +231,21 @@ namespace Anatawa12.AvatarOptimizer.APIInternal.VRCSDK
                             }
                             else
                             {
-                                BuildReport.LogWarning("ComponentInfos:VRCAvatarDescriptor:warning:NoMeshInEyelidsSkinnedMesh")
-                                    ?.WithContext(component);
+                                BuildLog.LogWarning("ComponentInfos:VRCAvatarDescriptor:warning:NoMeshInEyelidsSkinnedMesh",
+                                        component);
                             }
                         }
                         else
                         {
-                            BuildReport.LogWarning("ComponentInfos:VRCAvatarDescriptor:warning:NoEyelidsSkinnedMesh")
-                                ?.WithContext(component);
+                            BuildLog.LogWarning("ComponentInfos:VRCAvatarDescriptor:warning:NoEyelidsSkinnedMesh",
+                                    component);
                         }
                     }
                         break;
                     default:
-                        BuildReport.LogWarning("ComponentInfos:VRCAvatarDescriptor:warning:UnknownEyelidType", 
-                                component.customEyeLookSettings.eyelidType.ToString())
-                            ?.WithContext(component);
+                        BuildLog.LogWarning("ComponentInfos:VRCAvatarDescriptor:warning:UnknownEyelidType", 
+                                component.customEyeLookSettings.eyelidType.ToString(),
+                                component);
                         break;
                 }
             }
@@ -282,7 +279,7 @@ namespace Anatawa12.AvatarOptimizer.APIInternal.VRCSDK
                         }
                         
                         if (removed)
-                            BuildReport.LogFatal("ApplyObjectMapping:VRCAvatarDescriptor:eyelids BlendShape Removed");
+                            BuildLog.LogError("ApplyObjectMapping:VRCAvatarDescriptor:eyelids BlendShape Removed");
                     }
                         break;
                     default: