+++ Editor/APIInternal/ComponentInfos.VRM0.cs
@@ -2,7 +2,6 @@
 
 using System.Linq;
 using Anatawa12.AvatarOptimizer.API;
-using Anatawa12.AvatarOptimizer.ErrorReporting;
 using UnityEngine;
 using VRM;
 
@@ -152,7 +151,7 @@ namespace Anatawa12.AvatarOptimizer.APIInternal
                     else
                     {
                         mergedFirstPersonFlag = firstPersonFlags.Contains(FirstPersonFlag.Both) ? FirstPersonFlag.Both : FirstPersonFlag.Auto;
-                        BuildReport.LogWarning("MergeSkinnedMesh:warning:VRM:FirstPersonFlagsMismatch", mergedFirstPersonFlag.ToString());
+                        BuildLog.LogWarning("MergeSkinnedMesh:warning:VRM:FirstPersonFlagsMismatch", mergedFirstPersonFlag.ToString());
                     }
 
                     return new VRMFirstPerson.RendererFirstPersonFlags