+++ Internal/ErrorReporter/Editor/asssembly-info.cs
@@ -1,4 +1,0 @@
-using CustomLocalization4EditorExtension;
-
-// TODO: currently uses com.anatawa12.avatar-optimizer.editor but we need to migrate to own CL4EE instance later.
-[assembly:RedirectCL4EEInstance("com.anatawa12.avatar-optimizer.editor")]